fix: parent widget error
