+++ lib/screens/hosting/host_page_amenities.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../core/forms/host_top_menu_gradient.dart';
+import '../core/forms/top_menu_gradient.dart';
 import 'components/flow_builder_functions.dart';
 import 'components/host_big_iconbutton.dart';
 import 'components/host_bottombar.dart';
